fix No such file or directory error for new installs

+++ setup/webmail.sh
@@ -52,8 +52,10 @@ elif [[ "$UPDATE_KEY" != `cat /usr/local/lib/roundcubemail/version` ]]; then
 fi
 if [ $needs_update == 1 ]; then
   # if upgrading from 1.3.x, clear the temp_dir
-  if [ "$(cat /usr/local/lib/roundcubemail/version | cut -c1-3)" == '1.3' ]; then
-    find /var/tmp/roundcubemail/ -type f ! -name 'RCMTEMP*' -delete
+  if [ -f /usr/local/lib/roundcubemail/version ]; then
+    if [ "$(cat /usr/local/lib/roundcubemail/version | cut -c1-3)" == '1.3' ]; then
+      find /var/tmp/roundcubemail/ -type f ! -name 'RCMTEMP*' -delete
+    fi
   fi
 
 	# install roundcube
